// lib/api/generated/models/TagRelationCreateFieldMappingItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TagRelationCreateFieldMappingItem = {
  fieldName?: string;
  description?: string;
  dataType?: string;
  position?: number;
  isIdentity?: boolean;
  tagId?: number;
  isNotNull?: boolean;
};