// lib/api/generated/models/TagRelationFieldMappingDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TagRelationFieldMappingDto = {
  id?: number;
  tagRelationId?: number;
  fieldName?: string;
  description?: string;
  dataType?: string;
  position?: number;
  isIdentity?: boolean;
  tagId?: number;
  isNotNull?: boolean;
};