// lib/api/generated/models/TagRelationPermissionDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TagRelationPermissionDto = {
  id?: number;
  tagRelationId?: number;
  userId?: number;
  groupId?: number;
  canAccess?: boolean;
  canCreate?: boolean;
  canModify?: boolean;
  canDelete?: boolean;
  createBy?: number;
  createAt?: string;
  updateAt?: string;
  deleteAt?: string;
};