// lib/api/generated/models/TagRelationUpdateFieldMappingItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TagRelationUpdateFieldMappingItem = {
  id?: number;
  fieldName?: string;
  description?: string;
  dataType?: string;
  position?: number;
  isIdentity?: boolean;
  tagId?: number;
  isNotNull?: boolean;
};